third_party/blink/web_tests/platform/generic/fast/js/function-names-expected.txt

This test checks the names of all sorts of different functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new Function(' return 1; ').toString().replace(/[ \n]+/g, ' ') is 'function anonymous( ) { return 1; }'
PASS document.documentElement.onclick.toString().replace(/[ \n]+/g, ' ') is 'function onclick(event) { return 2; }'
PASS ''.constructor is String
PASS Boolean.toString() is 'function Boolean() { [native code] }'
PASS Date.toString() is 'function Date() { [native code] }'
PASS Error.toString() is 'function Error() { [native code] }'
PASS EvalError.toString() is 'function EvalError() { [native code] }'
PASS Function.toString() is 'function Function() { [native code] }'
PASS Number.toString() is 'function Number() { [native code] }'
PASS Object.toString() is 'function Object() { [native code] }'
PASS RangeError.toString() is 'function RangeError() { [native code] }'
PASS ReferenceError.toString() is 'function ReferenceError() { [native code] }'
PASS RegExp.toString() is 'function RegExp() { [native code] }'
PASS String.toString() is 'function String() { [native code] }'
PASS SyntaxError.toString() is 'function SyntaxError() { [native code] }'
PASS TypeError.toString() is 'function TypeError() { [native code] }'
PASS URIError.toString() is 'function URIError() { [native code] }'
PASS successfullyParsed is true

TEST COMPLETE
